move all the todo form to the todo form model
